pyproject.toml
[project]
name = "lv-wordrush-api"
version = "0.1.0"
description = "WordRush API backend"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.17",
    "djangorestframework==3.14.0",
    "drf-spectacular==0.27.2",
    "python-dotenv==1.0.1",
    "dj-database-url==1.0.0",
    "django-cors-headers==4.0.0",
    "gunicorn==20.1.0",
    "psycopg[binary]>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-django==4.7.0",
    "ruff==0.3.7",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["game/tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.lint.isort]
known-first-party = ["config", "game"]
